Revert "Adding main form buttons"
This reverts commit 51927f175c98e9b161edade4d5be7f08451cd11a.

+++ KWZP2022/KWZP2022/Parametr_czesc.cs
@@ -16,8 +16,8 @@ namespace KWZP2022
     {
         public int ID_czesc { get; set; }
         public int ID_rodzaj_parametr { get; set; }
-        public decimal Zakres_dol { get; set; }
-        public decimal Zakres_gora { get; set; }
+        public int Zakres_dol { get; set; }
+        public int Zakres_gora { get; set; }
     
         public virtual Czesc Czesc { get; set; }
         public virtual Rodzaj_parametr Rodzaj_parametr { get; set; }
